Refactoring model class name
