+++ services/product-service.js
@@ -1,4 +1,4 @@
-const Product = require('../models/product-model');
+const Product = require('../models/product');
 
 exports.createProduct = async function (product) {
     return await Product.create(product);
